pyproject.toml
[project]
name = "model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fuzzywuzzy>=0.18.0",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "python-levenshtein>=0.27.1",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "xgboost>=3.0.0",
]
